Use Long for a Content-Length header size.

+++ components/api/src/main/java/com/hotels/styx/api/HttpRequest.java
@@ -45,7 +45,7 @@ import static com.hotels.styx.api.HttpVersion.HTTP_1_1;
 import static com.hotels.styx.api.RequestCookie.decode;
 import static com.hotels.styx.api.RequestCookie.encode;
 import static io.netty.buffer.Unpooled.copiedBuffer;
-import static java.lang.Integer.parseInt;
+import static java.lang.Long.parseLong;
 import static java.util.Arrays.asList;
 import static java.util.Objects.requireNonNull;
 import static java.util.UUID.randomUUID;
@@ -730,7 +730,7 @@ public class HttpRequest implements HttpMessage {
 
         private static boolean isInteger(String contentLength) {
             try {
-                parseInt(contentLength);
+                parseLong(contentLength);
                 return true;
             } catch (NumberFormatException e) {
                 return false;
