+++ components/api/src/main/java/com/hotels/styx/api/LiveHttpResponse.java
@@ -39,7 +39,7 @@ import static com.hotels.styx.api.HttpVersion.httpVersion;
 import static com.hotels.styx.api.ResponseCookie.decode;
 import static com.hotels.styx.api.ResponseCookie.encode;
 import static io.netty.buffer.ByteBufUtil.getBytes;
-import static java.lang.Integer.parseInt;
+import static java.lang.Long.parseLong;
 import static java.util.Arrays.asList;
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toList;
@@ -871,7 +871,7 @@ public class LiveHttpResponse implements LiveHttpMessage {
 
         private static boolean isInteger(String contentLength) {
             try {
-                parseInt(contentLength);
+                parseLong(contentLength);
                 return true;
             } catch (NumberFormatException e) {
                 return false;
